pyproject.toml
[project]
name = "support-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Pavel14701",email = "pavelkutia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "dishka (>=1.5.2,<2.0.0)",
    "g4f (>=0.5.0.4,<0.6.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "curl-cffi (>=0.10.0,<0.11.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "faststream[rabbit] (>=0.5.39,<0.6.0)",
    "torch (>=2.6.0,<3.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
]

[tool.poetry]
packages = [{include = "support_assistant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
